Make Y-combinator work
This probably results in a few regressions in terms of outputting lambdas, but that's to be dealt with at a later time.

Turns out the factorial conversion isn't the main performance bottleneck; it runs just fine. `assertNumber` is the true resource hog - I think it's taking $O(N^3)$ time.

Either way it's way too late so here, take these changes @edwfeng.